+++ ving/index.mjs
@@ -5,7 +5,6 @@ import { addJob } from '#ving/jobs/queue.mjs';
 import { ouch } from '#ving/utils/ouch.mjs';
 import { sleep } from '#ving/utils/sleep.mjs';
 import { miniHash } from '#ving/utils/miniHash.mjs';
-//import { useKind } from '#ving/record/utils.mjs'
 import { useDB } from '#ving/drizzle/db.mjs'
 import { useRedis } from '#ving/redis.mjs'
 import { getConfig } from '#ving/config.mjs';
@@ -24,7 +23,6 @@ import { getConfig } from '#ving/config.mjs';
  * addJob() - the same as `addJob` from '#ving/jobs/queue.mjs'
  * ouch() - the same as `ouch` from '#ving/utils/ouch.mjs'
  * getConfig() - the sames `getConfig` from '#ving/config.mjs'
- * useKind() - the same as `useKind` from '#ving/record/VingRecord.mjs'
  * close() - closes the connections mysql, redis/BullMQ, and redis/cache
  */
 export default {
@@ -37,9 +35,8 @@ export default {
     addJob,
     ouch,
     getConfig,
-    //useKind,
     close: async () => {
-        await useDB().session.client.pool.end();
+        await useDB().$client.pool.end();
         await useRedis().quit();
     }
 }